Demo that processes can change syscall args after monitor checks
without a way for the monitor to check if the args have been
modified before submitting the "continue" response to the kernel.

Additionally, sending SIGSTOP to the process and only CONT'ing after
the "continue" response has been tested (manually) but it seems
processes terminate when a "continue" is sent while they're STOP'd.